finish bloc

+++ frontend_client/lib/model/request/product_request.dart
@@ -13,14 +13,14 @@ class ProductRequest extends Equatable {
     return ProductRequest(
       model: json['model'] as String?,
       description: json['description'] as String?,
-      price: json['price'] as int?,
+      price: json['price'] as double?,
       stock: json['stock'] as int?,
       categoryId: json['categoryId'] as int?,
     );
   }
   final String? model;
   final String? description;
-  final int? price;
+  final double? price;
   final int? stock;
   final int? categoryId;
